feat(web,sdf): default subscriptions UI
Adds the ability to mark attributes as default subscription sources, and
to display components grouped by whether they are subscribed to those
attributes.

Hidden behind the `default-subs` feature flag.

NOTE: Two parts of this are unfinished: it is not fully styled. And
there is a bug in the interaction with the default subscription where
the attribute subscription list will intercept the checkbox click and
trigger blur, focusing the attribute input. But I would like to merge
this behind the flag and follow up on finishing up the rest of it.

Co-authored-by: Zachary Hamm <zack@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>