WIP: make pinning, group by, and default subs one exclusive mode
This not only makes things WAY easier when we eventually move to
ExploreGrid and ExploreGridRow, but it also allows you to not have the
group by button remain disabled.

As a reminder, we need this to work with the URL query parameters.

Other note: no interaction on the header for the new "contentHeader"
other than the tags area (which is exactly how the subscriptions panel
plays out).

Co-authored-by: Zachary Hamm <zack@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>