feat(web): Add ability to groupBy incompatible components
When a component has been migrated from the old workspace structure to 
the new, a number of them are in an incompatible state. That means they 
have a socket connection that can’t be migrated

In larger workspaces they are difficult to find. So now we will 
determine if there are components in the change set that are 
incompatible and if show, allow you to groupBy those.

If you have an incompatible groupBy and change to a changeSet where are 
none, then it will remove that groupBy

https://jam.dev/c/b81edd7b-0402-4394-b77f-42e4248f324d