chore(sdf-api-test): Refactor mjolnir api test to use new v2 routes, add some helpers for use to expand and rewrite others
This is part 1 of 3. Once this merges, we refactor the other tests to use this pattern and stop using all of the old routes. After that, we can refactor the github actions pipeline that triggers these to support easier expansion without having to create and seed workspaces across environments. 

I also added an abandon all to the test that applies a change set as we've seen those workspaces grow in the number of open change sets due to various test failures.