fix(dal): if conn state is invalid, error instead of panicking
We're seeing panics in production in `txns`, which could only happen if
the connection state is "invalid". We're investigating why the state is
invalid after we try to run a transaction, but panicking is the wrong
thing to do here, no matter what. So let's produce an error instead of
panicking.