fix(si-mcp-server): Ensure MCP server prefers apply over approval
I have changed the `change-set-update-status` tool. Now if a user passes
`apply` then it’s going to determine if it needs to request approval or
not the for the change set. A user can specify that they don’t care 
about approvals using `force-apply`.

We have also updated the luminork endpoint so that if a user requests 
approval and the workspace hasn’t gotten approvals enabled, then it will
error