Fix left click on Map with retain state
This change fixes the load on the Map view after exiting a component
view. What happens is that we want to retain session state, just like
in the grid view using the "Explore.vue" watch. The problem is that the
Map has its own "selectedComponents" watcher, so it steamrolls over the
retain session state functionality. The early return added here should
avoid the situation.

Co-authored-by: Victor Bustamante <victor@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>