feat(dal): Allow management functions to directly set a resource for a component
This change removes the root/resource tree from the list of paths management functions are not allowed to update. We recently made a change that allowed refresh actions to run in a change set immediately after import management functions ran, however on considering expanding this behavior to Discovery management functions as well, it became obvious that all of this could be avoided by simply allowing the Discover/Import functions to write to root/resource as well as root/si and root/domain.  

This change enables us to update the management functions accordingly and remove the special action dispatching that was going on.