fix(web): Special case creating deleteComponents endpoint
Both erase and delete components endpoints are the same, we just pass a 
different parameter. 

This means that creating a delete and erase change set will generate the
same name. That was difficult to understand for users. So we are going 
to call it “remove components” as a middle ground