Pipe through component audit logs to history
This change pipes through component-specific audit
logs to the new component history panel in the new
review screen. This is a very minimal styling pass
mainly intended to put functionality first.

I forgot to co-author @johnrwatson on the commit
from the last PR (#7012) despite putting him in
the PR description, but here he is on this
commit!

Co-authored-by: John Watson <john@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>