fix(rebaser): When dispatching actions, ensure Edda knows the graph has been mutated again
We recently modified the Actions Job to no longer modify the graph to 
capture the action’s now-running state. We planned on exposing 
`Dispatched` as the user indicator that the action is in flight, however
because the dispatch happens after the rebase request has been processed
(and Edda has been informed), we need to again, tell Edda about the 
change to the graph that the Rebaser does directly in post processing.