feat(naxum): add optional status code to `JetstreamPostProcess` failures
This change updates the `OnFailure::call` method signature to include
`status: Option<StatusCode>`. As any `impl std::error::Error`'s have
already been converted down into a `Response`, this `OnFailure`
callback doesn't have access to the error which resulted in this
callback being fired. However, a `StatusCode` can often be as useful to
determine the *class* of the triggering error type and thus what needs
to happen in the function. Future work will use this in the Rebaser
codebase.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>