feat(edda): Additional Diff MVs and Secret Data
In order to correctly populate the Review screen, we need additional
pieces of information:
`ErasedComponents` provides a map of components that exist on head but do not exist in this change set. Because we replay changes to head on open change sets, any change to components on head will trigger this MV to build. If you’ve erased a component in this change set, but it still exists on head, we’ll provide the same ComponentDiff, with all data populated on head. 

`ActionDiffList` provides a list of Actions and their diff status: 
- If the state has been changed in this change set (for example, putting an action on hold)
- If the action is new in this change set
- If the action has been removed in this change set

`ComponentDiff` now includes information specific to Secrets so we can correctly show when changes have been made regarding secrets. Note: We aren’t diffing the actual encrypted secret data, but whenever the secret data changes, we always end up with a new secret so we can show when non-encrypted data changes (name/description) as well as show when subscriptions to secrets change

I also fixed a bug where we were using the wrong `DalContext` when computing the `ComponentDiff`