feat(web): Port Component Debug to new hotness
Viewing the debug information for a component is one of the few reasons
we still navigate to the old experience, so this ports it over, behind a
Flag, so it won’t show up for users.

Adds a new Panel to ComponentDetails for Debug, that when opened,fetches
debug data from a new V2 Route. I also removed the old debug route, as
well as the socket based autoconnect route which is no longer needed.

Features include: Searching across all of the debug data, and
highlighting the row where the data is found. I also kept some basic
socket info, mostly so we can still see which components have sockets
(even if they’re not in use).

Warning: `ComponentDebugPanel.vue` is particularly a Claude special :)