Add actions to review and icons expand modals
This change adds actions to the review screen. In addition to this
change, the actions pills themselves have been split into a new
component for both the review screen and grid tiles to use. This is an
intentional port of the "ActionsPanel" without additional styling
changes.

This change also adds icons to expand modals. This doesn't only add
icons to the actions modal in the review screen, but also to all other
expand modals, like qualifications in component details.

Signed-off-by: Nick Gerace <nick@systeminit.com>