Add component-generate-template MCP tool
This change adds a tool for generate templates from components. The
tool not only generates the template, but it also finds the resulting
schema. Along the way to making this change, the following was also
changed:

- Format all "si-mcp-server" files using the buck2 target "fix-format"
  (side note: we should be checking for this in CI)
- Add an additional example finding buck2 targets locally to the dev
  docs (this was used to find the "fix-format" command)
- Return the schema ID in addition to the existing data in the luminork
  route for generating a template
- Try to find the schema on the graph first before going to cached
  modules when finding schemas in luminork

The primary change was tested by creating a VPC with two subnets and
then generating a template from it. The template was then ran to confirm
that it worked.

Co-authored-by: Aaron Dernley <aaron@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>