fix(luminork): Allow a user to set a secret to a secret attribute
When a component is for a secret defining variant (i.e. AWS Credential) then we need to let a user to be able to easily set a secret directly on the secret Attribute Value. If we force a user to set that secret via the attributes API, then we would require a user to know the SecretId to use.

We will now allow a user to set a secret directly. This will not work for subscriptions and will NOT work for components that are not secret definiting. They will get a 422 if they try to do that