feat(dal): schema level actions, overlays
Adds the ability to add an ActionPrototype as an outgoing target of a
Schema instead of a Schema Variant. When picking the action to run, the
schema level action for a specific kind (Create, Refresh, Destroy,
Update) will be chosen instead of the schema variant level action.

Schema level action prototypes will also be made a child of a new
category node: Overlays. This will allow us to find "overlay" action
prototypes and to relate them to more than one schema.