Add the rebuildinator
This change adds the rebuildinator, which is a wrapper around the edda
client to build MVs for a given list of workspace and change set pairs.

If invalid IDs are provided, then the requests will sit on
"EDDA_REQUESTS" stream, but they are ultimately not harmful. The
rebuildinator intentionally does not pull in the dal in order to
minimize dependency usage. Thus, both the workspace ID and change set
ID are required (no practical way to determine the workspace from the
change set without the dal).

Co-authored-by: Aaron Dernley <aaron@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>