Add event for created funcs in luminork
This change adds a WsEvent for created and attached funcs in luminork.
This fixes an issue where authoring with the MCP server resulted in
not seeing the new funcs appear in the customization UI. Publishing the
event is now shared logic between sdf and luminork.

In addition to the above, this change adds a README for luminork.

Co-authored-by: Aaron Dernley <aaron@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>