Proceed without the previous version if unable to retrieve during deployment MV building
We apparently have some deployment-level MVs in the MvIndex where we are
missing the MV itself. While it's a problem that we were missing the
actual MV listed in the index, we're also in the process of building the
missing MV when this error happens. Now, if we encounter this, we'll log
a warning that it happened so we don't have this happening without any
visibility, but we'll also proceed as if it wasn't in the MvIndex to
start with to allow the build to continue and hopefully finish.