fix(web): Ensure toDelete on HEAD doesn't leak to change set review
When a component is marked as toDelete on HEAD and that is rebased to a change set, we ensure that we check the diffStatus before we show it. Without this check, we would get the component showing in the review screen for a change set. We don’t want that behaviour