fix(dal): cleanup graph after value clone to remove hanging APAs
When duplicating a component we remove the existing prototypes, but
those prototypes have attribute prototype arguments that are not removed
right away. These can lead to phantom "subscriptions" that have no
prototype (because the prototype has been removed).

Also removes a block of code that does nothing, since the prototype has
just been set to a component prototype it will always have an
source of "AttributeValue"