chore(sdf,luminork): skip loading snapshot when abandoning a change set
This change attempts to reduce the chances error or long waits when
abandoning a change set. Prior to this change the workspace snapshot for
the target change set was loaded via the
`DalContext::update_visibility_and_snapshot_to_visibility` method.
No code in either the SDF or Luminork API handler is using the loaded
workspace snapshot, and in the case of another bug, this load led to an
error which resulted in not being able to abandon a change set. Finally,
the workspace snapshot load can take some time if it is not already in
cache, and will take up to 10 seconds if the snapshot is no longer
available in the layer cache database (this is due to the current layer
cache spin lock retry logic).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>