fix(edda): Stop building CachedSchemaVariant MVs as we only ever want to return the latest variant for a builtin, which is the default variant
Currently, when modules are promoted as builtins, they only contain a single schema, and a single variant (the default variant, which is the latest).  When exposing builtins as uninstalled, we only ever want to show the latest variant (which is reflected as the default) as we don’t want users to install old builtins.

We decided to stop building these while working on supporting incremental builds of the Deployment MVs, and at that time, we found that if we simply stop building the CachedSchemaVariantMV alltogether, it brings the total build time down to 16s (locally) for all current builtins. 

Given this, we likely don’t need to support incremental builds for some time, so we’ll ship this as a stand alone improvement and wait for our needs to grow again to revist incremental builds