feat(web): use bulk queries for workspace patches
This changes the workspace patch handling to insert all new patches in
bulk, and to be sure not to apply patches if the atom already exists.

This should make large patch batches much faster than inserting atoms
one query at a time. It also ensures we do not run any sqlite3 ops in
parallel, which can slow sqlite3 down.

Also fixes a bug in handleIndexMvPatch where we were not correctly checking
for NOROW.