feat(web): use bulk queries for workspace patches
This changes the workspace patch handling to insert all new patches in
bulk, and to be sure not to apply patches if the atom already exists.

This should make large patch batches much faster than inserting atoms
one query at a time. It also ensures we do not run any sqlite3 ops in
parallel, which can slow sqlite3 down.

Fixes empty patch handling to ensure indexes are not dropped, which
broke handleIndexMvPatch.

Test case update to catch & cover error we are seeing in prod

Co-Authored-By: John Obelenus <jobelenus@systeminit.com>