feat(web): Skip onboarding flow for invited users
If a user has onboard to a workspace and then invites another user and that user is new to SI, we should skip the onboarding flow for that invited user.

The way we determine if the workspace has been onboarded is that there are components on HEAD. These components would be the credential and the region as part of the onboarding

If a user deletes them, then it’s fine for an invited user to go through onboarding as the aim of the game is to make that workspace ready for action!