feat(luminork): default to overlay actions and mgmt funcs for builtins
When an action or management function is created in luminork, default to
creating an "overlay" (schema parented) function and prototype instead
of a variant-level function. This will allow the custom functions to
persist between upgrades, and does not require locking the variant.

In order to make this work with our materialized views, adds a
post-processing step to change detection that adds SchemaVariant changes
to the change set to force the SchemaVariant related materialized views
to rebuild so that they reflect these new functions.

Tests:

1. Create a new management function for a builtin via luminork. Observe
   that it is available in the management functions dropdown in the UI.
   Try to run it.

2. See that the function exists in the list of functions when you fetch
   the schema variant in luminork.

3. Create an action for a builtin via luminork. Confirm you cannot
   create a second action of that type. Add a component and enqueue that
   action. Apply to head. See that the action that ran was the new one,
   and not the default one for that variant.

4. See that the action function exists in the func list when you fetch
   the variant via luminork.

Additional tests:

  Use luminork to create a new create action for one change set, and a
  new create action in another change set, both for the same schema.
  Apply one of the change sets to head. Ensure that create actions work
  when applying the second change set after the first has been applied.