feat: leaf overlay functions
Introduces LeafPrototypes, which are direct children of schemas. They
point to a LeafKind function (Qualification or CodeGen) and are executed
in the dependent values update. Since they are not part of the schema
variant, they have to create their destination attribute values on the
fly (the key/value pair for the qualificaton or code maps).