chore(sdf): Re-enqueue the deployment Mv build on migration
We had disabled enqueuing the Edda task that would build the Deployment MVs, when the builtins are cached during migration as the task was taking 4ish hours. Now that this task completes rather quickly, let’s do this on migration again. This mostly improves the local dev experience so we don’t have to manually enqueue it to see the builtins available for component creation on startup.