Do not load mgmt func runs for invalid components
If you put a component ID in the URL that does not exist on the current
change set, you will see the "no component found" banner on the details
page. However, the latest management func runs route will still be
called. This changes blocks from it being called. This change also
converts two error in sdf and luminork, respectively, from 500 to 404
based on user actions made as a result of the original bug.

Signed-off-by: Nick Gerace <nick@systeminit.com>