Prevent incomplete subscriptions
This change prevents incomplete subscriptions in edge case scenarios by
removing new node updates for attribute prototype arguments
conditionally. This condition is triggered when there is a set
difference between the set of all destinations that no longer exist in
the graph (i.e. the "incomplete" subscriptions) and the set of all new
node updates.

This change fixes a bug where the erasure of a subscription component
(not the subscriber) in one change set would result in subscriber
components containing incomplete subscriptions when HEAD also contained
changes to the subscription itself. This case is tested in an
integration test.

Co-authored-by: Zachary Hamm <zack@systeminit.com>
Co-authored-by: Brit Myers <brit@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>