Prevent orphaned update actions
This change prevents orphaned update actions in a non-HEAD change set
by adding a correct transform. The transform ensures that the action has
a corresponding component in either the graph or the list of updates.

This change also refactors the correct transforms in the action node
weight module to separate scenarios where the node exists on the graph
or not. The split graph implementation remains knowingly untested
(checked with @zacharyhamm), but has been ported.

Finally, this change uses an "is subset" check rather than a
"difference" check for both the attribute prototype argument node
(PR #7571) to ensure we do not remove updates when "valid additions"
occur. This is the same motion used for action nodes in this PR.

Co-authored-by: Zachary Hamm <zack@systeminit.com>
Co-authored-by: John Watson <john@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>