feat(web): Filter actions by originating change set in actions view list
When viewing a change set, only show actions that originated in that specific change set. Previously, all actions were displayed regardless of which change set they originated from, which could be confusing for users working in different change sets.

Why filter this in the front end, currently the MV reads the data from the graph so the graph has all of the actions from HEAD that are replayed to the change set. Changing the graph here is way more risky, we can always remove the filter at a later date if we decide to do this a different way

Added 4 new test cases covering change set filtering scenarios:
- HEAD shows all actions regardless of origin
- Change sets only show actions from current change set
- No actions shown when all are from different change sets
- Correct filtering with multiple components and change sets