fix(dal): ensure deletes and upgrades cannot break components with subscriptions
Suppose you have a component A that is the source for a value, on HEAD.
And, in another change set, you have B, which has an outgoing
ValueSubscription edge to the root attribute value of A. If you either:
upgrade A to a new schema variant and apply to HEAD, or delete A and
apply to HEAD, we will not remove the root attribute value or the
incoming ValueSubscription edges for that deleted component without this
correction. The result is that the component B will be broken in the
change set.

This correction ensures that any time we either delete a component, or
detect that its schema variant is being upgraded, we esnure the root
attribute value is removed and we ensure that the attribute prototype
arguments that are the source for the ValueSubscription edges are also
removed, so that the subscribing (destination) component does not have
broken attribute values.