Split frigg entry errors and add index span fields
This change splits entry errors in frigg into three discrete types.
Two of them correspond to fetching change set indices, which are
currently within "debug-level" spans. This should help disambiguate
the types of upstream NATS client entry API errors with JetStream KV.

In addition, this change adds fields for workspace and change set in the
spans for rebuilding the index in SDF. The rebuild functions are now
located next to their routes too.

Signed-off-by: Nick Gerace <nick@systeminit.com>