Clarify scenarios for getting a change set index
This change clarifies the different scenarios that can be encountered
when getting a change set index. Before this change, "None" meant
multiple different things, which could be interpreted differently by
different callers. Why not use the "debug" logging already there and
add more? We do not want to spam useless logs and having the caller
explicitly decide what to do is much easier to debug and trace.

This change also changes luminork's search API to handle missing or
invalid indices with a 412 status code and the raw errors from Frigg
rather than a conversion.

Signed-off-by: Nick Gerace <nick@systeminit.com>